Add one button shortcut to change duration in order
This feature is intended for extending expired orders in a notification.
This is the easiest way to to this for now. I'm not sure if it's useful
to decrease duration except for hiding order from order book, so maybe
it would make more sense if editing duration could only extend it and
there was another button to archive order to hide it from order book
without deleting it, preserving duration.

Signed-off-by: alfred richardsn <rchrdsn@protonmail.ch>