Add tickers whitelist
In order creation users are now required to choose from the whitelists of
currencies and gateways of the chosen currency. If user sends a currency
or gateway that's not whitelisted, they can request whitelisting. If two
unique requests are received, it's sent to support chat with a count of
unique requests.

Signed-off-by: alfred richardsn <rchrdsn@protonmail.ch>