Change translation message IDs to keys
When I first added i18n, I thought mostly of my own convenience, and
it's easier for me to work with untranslated English strings as
translation message IDs, because I can see what exactly bot sends in the
code and it's much easier for me to translate it in Russian.

But through a discussion with Fincubator members I decided to change
untranslated English strings to arbitrary keys as message IDs. It would
provide a little more challenge for developers, but it would give less
headache to translators and editors:

        * Nobody has to touch code and regenerate .po files to change
          wording in English, which is a good idea if we let non-developers
          to do that.
        * If wording in English has changed, translations don't break.

Of course, if the meaning of English message has changed, translations
should intentionally be broken, but it's not always required, so in that
case we would just change key name.

The gettext keyword of strings has also changed from _ to i18n. Using
two names for the same object just to distinguish its meaning was kind
of confusing anyway, considering _ as a variable has its own meaning in
Python which now can be utilized.

To generate .pot file via Babel:

        pybabel extract src/* -k i18n -o locale/bot.pot --add-location=file

I'll try to find a proper way to assign these options to the project.

Signed-off-by: alfred richardsn <rchrdsn@protonmail.ch>