Simplify config module
Access to configuration values is now done through an instance of lazy
interface to typed environment variables, which allows not to update
this module every time a new variable is added (unless it must have a
default value). From now on, addition of new variables will be done
primarily just by appending them to .env.example.

Signed-off-by: alfred richardsn <rchrdsn@protonmail.ch>