Add cyberway support to escrow
Blockchain base class was changed to fit more blockchains. In
particular, part of it responsible for streaming new trasactions was
moved to another class, as it was decided to handle transaction check of
cyberway (which doesn't support streaming) by pressing on button instead
of starting check immediately to reduce node load.

As TellerBot uses node which is not public, node URLs were moved to
secrets.

Signed-off-by: alfred richardsn <rchrdsn@protonmail.ch>